LinkTranslator gets a separate file
