+++ src/realm/table.cpp
@@ -333,73 +333,6 @@ void LinkChain::add(ColKey ck)
     m_link_cols.push_back(ck);
 }
 
-LinkTranslator::LinkTranslator(Obj origin, ColKey origin_col_key)
-    : m_origin_obj(origin)
-    , m_origin_col_key(origin_col_key)
-{
-}
-
-void LinkTranslator::run()
-{
-    ColumnAttrMask attr = m_origin_col_key.get_attrs();
-    if (attr.test(col_attr_List)) {
-        if (m_origin_col_key.get_type() == col_type_LinkList) {
-            LnkLst link_list = m_origin_obj.get_linklist(m_origin_col_key);
-            on_list_of_links(link_list);
-        }
-        else if (m_origin_col_key.get_type() == col_type_Mixed) {
-            Lst<Mixed> list = m_origin_obj.get_list<Mixed>(m_origin_col_key);
-            on_list_of_mixed(list);
-        }
-        else if (m_origin_col_key.get_type() == col_type_TypedLink) {
-            Lst<ObjLink> list = m_origin_obj.get_list<ObjLink>(m_origin_col_key);
-            on_list_of_typedlink(list);
-        }
-        else {
-            throw std::runtime_error(
-                util::format("LinkTranslator unhandled list type: %1", m_origin_col_key.get_type()));
-        }
-    }
-    else if (attr.test(col_attr_Set)) {
-        if (m_origin_col_key.get_type() == col_type_Link) {
-            LnkSet set = m_origin_obj.get_linkset(m_origin_col_key);
-            on_set_of_links(set);
-        }
-        else if (m_origin_col_key.get_type() == col_type_Mixed) {
-            Set<Mixed> set = m_origin_obj.get_set<Mixed>(m_origin_col_key);
-            on_set_of_mixed(set);
-        }
-        else if (m_origin_col_key.get_type() == col_type_TypedLink) {
-            Set<ObjLink> set = m_origin_obj.get_set<ObjLink>(m_origin_col_key);
-            on_set_of_typedlink(set);
-        }
-        else {
-            throw std::runtime_error(
-                util::format("LinkTranslator unhandled set type: %1", m_origin_col_key.get_type()));
-        }
-    }
-    else if (attr.test(col_attr_Dictionary)) {
-        auto dict = m_origin_obj.get_dictionary(m_origin_col_key);
-        on_dictionary(dict);
-    }
-    else {
-        REALM_ASSERT_EX(!m_origin_col_key.is_collection(), m_origin_col_key);
-        if (m_origin_col_key.get_type() == col_type_Link) {
-            on_link_property(m_origin_col_key);
-        }
-        else if (m_origin_col_key.get_type() == col_type_Mixed) {
-            on_mixed_property(m_origin_col_key);
-        }
-        else if (m_origin_col_key.get_type() == col_type_TypedLink) {
-            on_typedlink_property(m_origin_col_key);
-        }
-        else {
-            throw std::runtime_error(
-                util::format("LinkTranslator unhandled property type: %1", m_origin_col_key.get_type()));
-        }
-    }
-}
-
 // -- Table ---------------------------------------------------------------------------------
 
 ColKey Table::add_column(DataType type, StringData name, bool nullable)
